Put guard around test dropping tables so it doesn't try to drop one that doesn't exists

+++ src/test/scala/app/models/TaskDAOSpec.scala
@@ -17,7 +17,7 @@ class TaskDAOSpec extends WordSpec with Matchers with BeforeAndAfter {
       driver = C.dbDriver
     ).
       createSession()
-    dropTable()
+    if(tables.size > 0) dropTable()
   }
 
   "createTable" should {
